pyproject.toml
[tool.poetry]
name = "Complete FastAPI REST course + AWS + Emails + Payments"
version = "0.1.0"
description = ""
authors = ["Edilson Silva <eddysilva1994@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.35"
databases = "^0.5.5"
asyncpg = "^0.25.0"
psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyJWT = "^2.3.0"
email-validator = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"
alembic = "^1.7.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
